big refactor to clean up the household member section
