+++ src/Components/EnergyCalculator/Steps/HouseholdMemberForm.tsx
@@ -1,6 +1,6 @@
 import { FormattedMessage, useIntl } from 'react-intl';
 import QuestionHeader from '../../QuestionComponents/QuestionHeader';
-import HHMSummaryCards from '../../Steps/HouseholdMembers/HHMSummaryCards';
+import HouseholdMemberSummaryCards from '../../Steps/HouseholdMembers/components/HouseholdMemberSummaryCards';
 import { useNavigate, useParams } from 'react-router-dom';
 import { Context } from '../../Wrapper/Wrapper';
 import { ReactNode, useContext, useEffect, useMemo } from 'react';
@@ -26,7 +26,7 @@ import { useStepNumber } from '../../../Assets/stepDirectory';
 import * as z from 'zod';
 import { Controller, SubmitHandler, useFieldArray } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { MONTHS } from '../../Steps/HouseholdMembers/MONTHS';
+import { MONTHS } from '../../Steps/HouseholdMembers/utils/data';
 import PrevAndContinueButtons from '../../PrevAndContinueButtons/PrevAndContinueButtons';
 import ErrorMessageWrapper from '../../ErrorMessage/ErrorMessageWrapper';
 import MultiSelectTiles from '../../OptionCardGroup/MultiSelectTiles';
@@ -46,14 +46,14 @@ import {
   renderIncomeFrequencyHelperText,
   renderHoursWorkedHelperText,
   renderIncomeAmountHelperText,
-} from '../../Steps/HouseholdMembers/HelperTextFunctions';
+} from '../../Steps/HouseholdMembers/utils/validation';
 import { DOLLARS, handleNumbersOnly, numberInputProps, NUM_PAD_PROPS } from '../../../Assets/numInputHelpers';
 import useScreenApi from '../../../Assets/updateScreen';
 import { QUESTION_TITLES } from '../../../Assets/pageTitleTags';
 import { getCurrentMonthYear, YEARS, MAX_AGE } from '../../../Assets/age';
 import { useAgeCalculation } from '../../AgeCalculation/useAgeCalculation';
 import { determineDefaultIncomeByAge } from '../../AgeCalculation/AgeCalculation';
-import '../../../Components/Steps/HouseholdMembers/PersonIncomeBlock.css';
+import '../../Steps/HouseholdMembers/styles/PersonIncomeBlock.css';
 import { useShouldRedirectToConfirmation } from '../../QuestionComponents/questionHooks';
 import useStepForm from '../../Steps/stepForm';
 
@@ -948,7 +948,7 @@ const ECHouseholdMemberForm = () => {
           />
         )}
       </QuestionHeader>
-      <HHMSummaryCards
+      <HouseholdMemberSummaryCards
         activeMemberData={{
           ...getValues(),
           id: formData.householdData[currentMemberIndex]?.id ?? crypto.randomUUID(),
