+++ src/Components/Steps/HouseholdMembers/index.ts
@@ -1,0 +1,11 @@
+/**
+ * HouseholdMembers Module
+ *
+ * This module handles the household member data collection flow.
+ * Main entry point exports the router component.
+ */
+
+export { default } from './components/HouseholdMemberRouter';
+export { default as HouseholdMemberRouter } from './components/HouseholdMemberRouter';
+export { default as HouseholdMemberForm } from './components/HouseholdMemberForm';
+export { default as HouseholdMemberBasicInfoPage } from './components/HouseholdMemberBasicInfoPage';
